cleanup!: reduce service override count by merging them
