+++ rollup/rollup.config.ts
@@ -216,9 +216,7 @@ const input = {
   configuration: './src/service-override/configuration.ts',
   keybindings: './src/service-override/keybindings.ts',
   textmate: './src/service-override/textmate.ts',
-  languageConfiguration: './src/service-override/languageConfiguration.ts',
   theme: './src/service-override/theme.ts',
-  tokenClassification: './src/service-override/tokenClassification.ts',
   snippets: './src/service-override/snippets.ts',
   languages: './src/service-override/languages.ts',
   audioCue: './src/service-override/audioCue.ts',
