+++ rollup/rollup.types.config.ts
@@ -29,9 +29,7 @@ export default rollup.defineConfig([
   './dist/types/src/service-override/configuration.d.ts',
   './dist/types/src/service-override/keybindings.d.ts',
   './dist/types/src/service-override/textmate.d.ts',
-  './dist/types/src/service-override/languageConfiguration.d.ts',
   './dist/types/src/service-override/theme.d.ts',
-  './dist/types/src/service-override/tokenClassification.d.ts',
   './dist/types/src/service-override/snippets.d.ts',
   './dist/types/src/service-override/languages.d.ts',
   './dist/types/src/service-override/audioCue.d.ts',
