--- src/service-override/tokenClassification.ts
@@ -1,21 +1,0 @@
-import '../polyfill'
-import '../vscode-services/missing-services'
-import { IEditorOverrideServices } from 'vs/editor/standalone/browser/standaloneServices'
-import { TokenClassificationExtensionPoints } from 'vs/workbench/services/themes/common/tokenClassificationExtensionPoint'
-import { IWorkbenchContribution, IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions } from 'vs/workbench/common/contributions'
-import { Registry } from 'vs/platform/registry/common/platform'
-import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle'
-import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation'
-export class ExtensionPoints implements IWorkbenchContribution {
-  constructor (
-    @IInstantiationService private readonly instantiationService: IInstantiationService
-  ) {
-    this.instantiationService.createInstance(TokenClassificationExtensionPoints)
-  }
-}
-
-Registry.as<IWorkbenchContributionsRegistry>(WorkbenchExtensions.Workbench).registerWorkbenchContribution(ExtensionPoints, LifecyclePhase.Starting)
-
-export default function getServiceOverride (): IEditorOverrideServices {
-  return {}
-}
