+++ tsconfig.types.json
@@ -66,9 +66,7 @@
 		"src/service-override/configuration.ts",
 		"src/service-override/keybindings.ts",
 		"src/service-override/textmate.ts",
-		"src/service-override/languageConfiguration.ts",
 		"src/service-override/theme.ts",
-		"src/service-override/tokenClassification.ts",
 		"src/service-override/snippets.ts",
 		"src/service-override/languages.ts",
 		"src/service-override/audioCue.ts",
